@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <AssemblyName>opalopc</AssemblyName>
     <AssemblyVersion>1.0.0.0</AssemblyVersion>
     <OutputType>Exe</OutputType>
