@@ -7,28 +7,22 @@ using Util;
 
 namespace Controller
 {
-    public class DiscoveryController
+    public interface IDiscoveryController
     {
-        private readonly ILogger _logger;
-        private readonly CancellationToken? _token;
-        private readonly IDiscoveryUtil _discoveryUtil;
+        ICollection<Target> DiscoverTargets(ICollection<Uri> discoveryUris);
+    }
 
-
-        public DiscoveryController(ILogger logger, IDiscoveryUtil discoveryUtil, CancellationToken? token = null)
-        {
-            _logger = logger;
-            _token = token;
-            _discoveryUtil = discoveryUtil;
-        }
+    public class DiscoveryController(ILogger<DiscoveryController> logger, IDiscoveryUtil discoveryUtil, ITaskUtil taskUtil) : IDiscoveryController
+    {
 
         public ICollection<Target> DiscoverTargets(ICollection<Uri> discoveryUris)
         {
-            _logger.LogDebug("{Message}", $"Starting Discovery with {discoveryUris.Count} URIs");
+            logger.LogDebug("{Message}", $"Starting Discovery with {discoveryUris.Count} URIs");
 
             ICollection<Target> targets = new List<Target>();
             foreach (Uri uri in discoveryUris)
             {
-                TaskUtil.CheckForCancellation(_token);
+                taskUtil.CheckForCancellation();
                 targets = targets.Concat(DiscoverTargets(uri)).ToList();
             }
 
@@ -36,7 +30,7 @@ namespace Controller
         }
 
         // Given discoveryUri, discover all applications
-        private ICollection<Target> DiscoverTargets(Uri discoveryUri)
+        private List<Target> DiscoverTargets(Uri discoveryUri)
         {
 
             // https://reference.opcfoundation.org/Core/Part4/v105/docs/
@@ -55,15 +49,15 @@ namespace Controller
             // https://reference.opcfoundation.org/Core/Part4/v104/docs/5.4.2
             // ask the server for all servers it knows about
 
-            ICollection<Target> targets = new List<Target>();
+            List<Target> targets = [];
 
 
-            _logger.LogDebug("{Message}", $"Discovering applications in {discoveryUri}");
+            logger.LogDebug("{Message}", $"Discovering applications in {discoveryUri}");
 
             if (discoveryUri.ToString().Contains("https://"))
             {
                 string msg = $"Https is not supported: {discoveryUri}";
-                _logger.LogError("{Message}", msg);
+                logger.LogError("{Message}", msg);
                 return targets;
             }
 
@@ -75,7 +69,7 @@ namespace Controller
             }
             catch (Exception)
             {
-                _logger.LogWarning("{Message}", $"Skipping {discoveryUri} because of preceding errors");
+                logger.LogWarning("{Message}", $"Skipping {discoveryUri} because of preceding errors");
                 return targets;
             }
 
@@ -87,15 +81,15 @@ namespace Controller
             }
             catch (ServiceResultException)
             {
-                _logger.LogWarning("{Message}", $"Skipping {discoveryUri} because of preceding errors");
+                logger.LogWarning("{Message}", $"Skipping {discoveryUri} because of preceding errors");
                 return targets;
             }
 
-            _logger.LogDebug("{Message}", $"Discovered {adc.Count} applications");
+            logger.LogDebug("{Message}", $"Discovered {adc.Count} applications");
 
             foreach (ApplicationDescription ad in adc)
             {
-                TaskUtil.CheckForCancellation(_token);
+                taskUtil.CheckForCancellation();
                 targets.Add(DiscoverEndpoints(ad));
             }
 
@@ -108,7 +102,7 @@ namespace Controller
             try
             {
                 Uri uri = Utils.ParseUri(uriString) ?? throw new UriFormatException();
-                IPAddress[] addresses = _discoveryUtil.ResolveIPv4Addresses(uri.Host);
+                IPAddress[] addresses = discoveryUtil.ResolveIPv4Addresses(uri.Host);
                 string ip = addresses.First().ToString();
                 return uri.OriginalString.Replace(uri.Host, ip);
             }
@@ -118,17 +112,17 @@ namespace Controller
                 if (ex is UriFormatException)
                 {
                     msg = $"Invalid Uri format {uriString}";
-                    _logger.LogError("{Message}", msg);
+                    logger.LogError("{Message}", msg);
                 }
                 else if (ex is SocketException || ex is ArgumentException)
                 {
                     msg = $"Unable to resolve hostname {uriString}";
-                    _logger.LogError("{Message}", msg);
+                    logger.LogError("{Message}", msg);
                 }
                 else if (ex is ArgumentNullException || ex is ArgumentOutOfRangeException)
                 {
                     msg = $"Hostname or address of {uriString} is invalid";
-                    _logger.LogError("{Message}", msg);
+                    logger.LogError("{Message}", msg);
                 }
                 else
                 {
@@ -145,12 +139,12 @@ namespace Controller
 
             try
             {
-                adc = _discoveryUtil.DiscoverApplications(uri);
+                adc = discoveryUtil.DiscoverApplications(uri);
             }
             catch (Opc.Ua.ServiceResultException e)
             {
                 string msg = $"Cannot connect to discovery URI {uri}: {e}";
-                _logger.LogError("{Message}", msg);
+                logger.LogError("{Message}", msg);
                 throw;
             }
 
@@ -166,8 +160,8 @@ namespace Controller
 
                 // https://reference.opcfoundation.org/Core/Part4/v104/docs/5.4.4
                 // ask each discoveryUrl for endpoints
-                _logger.LogDebug("{Message}", $"Discovering endpoints for {applicationDescription.ApplicationName} ({applicationDescription.ProductUri})");
-                _logger.LogTrace("{Message}", $"Using DiscoveryUrl {s}");
+                logger.LogDebug("{Message}", $"Discovering endpoints for {applicationDescription.ApplicationName} ({applicationDescription.ProductUri})");
+                logger.LogTrace("{Message}", $"Using DiscoveryUrl {s}");
 
                 string s_by_ip;
                 try
@@ -176,7 +170,7 @@ namespace Controller
                 }
                 catch (Exception e)
                 {
-                    Server server = new(s, new EndpointDescriptionCollection());
+                    Server server = new(s, []);
                     server.AddError(new Error(e.Message));
 
                     target.AddServer(server);
@@ -186,9 +180,9 @@ namespace Controller
                 if (s_by_ip.Contains("https://"))
                 {
                     string msg = $"Https is not supported: {s_by_ip}";
-                    _logger.LogWarning("{Message}", msg);
+                    logger.LogWarning("{Message}", msg);
 
-                    Server server = new(s, new EndpointDescriptionCollection());
+                    Server server = new(s, []);
                     server.AddError(new Error(msg));
 
                     target.AddServer(server);
@@ -199,14 +193,14 @@ namespace Controller
 
                 try
                 {
-                    edc = _discoveryUtil.DiscoverEndpoints(new Uri(s_by_ip));
+                    edc = discoveryUtil.DiscoverEndpoints(new Uri(s_by_ip));
                 }
                 catch (Opc.Ua.ServiceResultException e)
                 {
                     string msg = $"Cannot connect to discovery URI {s_by_ip}: {e}";
-                    _logger.LogWarning("{Message}", msg);
+                    logger.LogWarning("{Message}", msg);
 
-                    Server server = new(s, new EndpointDescriptionCollection());
+                    Server server = new(s, []);
                     server.AddError(new Error(msg));
                     target.AddServer(server);
 
@@ -216,7 +210,7 @@ namespace Controller
                 // remove all that contain https scheme
                 edc.RemoveAll(e => e.EndpointUrl.Contains("https://"));
 
-                _logger.LogDebug("{Message}", $"Discovered {edc.Count} endpoints");
+                logger.LogDebug("{Message}", $"Discovered {edc.Count} endpoints");
 
                 target.AddServer(new Server(s, edc));
             }
