@@ -1,100 +1,86 @@
 using Microsoft.Extensions.Logging;
 using Model;
+using Org.BouncyCastle.Tsp;
 using Plugin;
 using Util;
 using View;
 
 namespace Controller
 {
-    public class ScanController
+    public interface IScanController
     {
-        readonly ILogger _logger;
-        readonly ICollection<Uri> _discoveryUris;
-        readonly Stream _reportOutputStream;
-        readonly string _commandLine;
-        readonly CancellationToken? _token;
-        readonly AuthenticationData _authenticationData;
+        void Scan(ICollection<Uri> discoveryUris, string commandLine, AuthenticationData authenticationData, Stream outputStream);
+    }
 
-        public ScanController(ILogger logger, ICollection<Uri> discoveryUris, Stream reportOutputStream, string commandLine, AuthenticationData authenticationData, CancellationToken? token = null)
+    public class ScanController(ILogger<ScanController> logger, IReportController reportController, IDiscoveryController discoveryController, ISecurityTestController securityTestController, ITaskUtil taskUtil) : IScanController
+    {
+        public void Scan(ICollection<Uri> discoveryUris, string commandLine, AuthenticationData authenticationData, Stream outputStream)
         {
-            _logger = logger;
-            _discoveryUris = discoveryUris;
-            _reportOutputStream = reportOutputStream;
-            _commandLine = commandLine;
-            _token = token;
-            _authenticationData = authenticationData;
-        }
-
-        public void Scan()
-        {
-            TelemetryUtil.TrackEvent("Scan started", GetScanProperties());
+            TelemetryUtil.TrackEvent("Scan started", GetScanProperties(discoveryUris, authenticationData));
 
             DateTime start = DateTime.Now;
-            _logger.LogInformation("{Message}", $"Starting OpalOPC {Util.VersionUtil.AppAssemblyVersion} ( https://opalopc.com )");
+            logger.LogInformation("{Message}", $"Starting OpalOPC {Util.VersionUtil.AppAssemblyVersion} ( https://opalopc.com )");
 
-            if (_discoveryUris.Count == 0)
+            if (discoveryUris.Count == 0)
             {
-                _logger.LogWarning("{Message}", "No targets were specified, so 0 applications will be scanned.");
+                logger.LogWarning("{Message}", "No targets were specified, so 0 applications will be scanned.");
             }
 
-            IReporter reporter = new Reporter(_reportOutputStream);
+            taskUtil.CheckForCancellation();
 
-            TaskUtil.CheckForCancellation(_token);
+            ICollection<Target> targets = discoveryController.DiscoverTargets(discoveryUris);
 
-            DiscoveryController discoveryController = new(_logger, new DiscoveryUtil(), _token);
-            ICollection<Target> targets = discoveryController.DiscoverTargets(_discoveryUris);
-
-            TaskUtil.CheckForCancellation(_token);
+            taskUtil.CheckForCancellation();
 
             // Initialize security testing plugins
             ICollection<IPlugin> securityTestPlugins = new List<IPlugin> {
-            new SecurityModeInvalidPlugin(_logger),
-            new SecurityModeNonePlugin(_logger),
+            new SecurityModeInvalidPlugin(logger),
+            new SecurityModeNonePlugin(logger),
 
-            new SecurityPolicyBasic128Rsa15Plugin(_logger),
-            new SecurityPolicyBasic256Plugin(_logger),
-            new SecurityPolicyNonePlugin(_logger),
+            new SecurityPolicyBasic128Rsa15Plugin(logger),
+            new SecurityPolicyBasic256Plugin(logger),
+            new SecurityPolicyNonePlugin(logger),
 
-            new AnonymousAuthenticationPlugin(_logger, _authenticationData),
-            new SelfSignedCertificatePlugin(_logger),
+            new AnonymousAuthenticationPlugin(logger, authenticationData),
+            new SelfSignedCertificatePlugin(logger),
 
-            new ProvidedCredentialsPlugin(_logger, _authenticationData),
-            new CommonCredentialsPlugin(_logger, _authenticationData),
-            new BruteForcePlugin(_logger, _authenticationData),
-            new RBACNotSupportedPlugin(_logger),
-            new AuditingDisabledPlugin(_logger),
+            new ProvidedCredentialsPlugin(logger, authenticationData),
+            new CommonCredentialsPlugin(logger, authenticationData),
+            new BruteForcePlugin(logger, authenticationData),
+            new RBACNotSupportedPlugin(logger),
+            new AuditingDisabledPlugin(logger),
         };
 
-            SecurityTestController securityTestController = new(_logger, securityTestPlugins, _token);
-            ICollection<Target> testedTargets = securityTestController.TestTargetSecurity(targets);
+            ICollection<Target> testedTargets = securityTestController.TestTargetSecurity(securityTestPlugins, targets);
 
-            TaskUtil.CheckForCancellation(_token);
+            taskUtil.CheckForCancellation();
 
             DateTime end = DateTime.Now;
 
             TimeSpan ts = (end - start);
-            string runStatus = $"OpalOPC done: {_discoveryUris.Count} Discovery URLs ({targets.Count} applications found) scanned in {Math.Round(ts.TotalSeconds, 2)} seconds";
-            _logger.LogInformation("{Message}", runStatus);
+            string runStatus = $"OpalOPC done: {discoveryUris.Count} Discovery URLs ({targets.Count} applications found) scanned in {Math.Round(ts.TotalSeconds, 2)} seconds";
+            logger.LogInformation("{Message}", runStatus);
 
-            ReportController reportController = new(_logger, reporter, testedTargets, start, end, _commandLine, runStatus);
-            reportController.WriteReport();
 
-            TelemetryUtil.TrackEvent("Scan finished", GetScanResultProperties(reportController.report, ts));
+            Report report = reportController.GenerateReport(testedTargets, start, end, commandLine, runStatus);
+            reportController.WriteReport(report, outputStream);
+
+            TelemetryUtil.TrackEvent("Scan finished", GetScanResultProperties(report, ts, discoveryUris, authenticationData));
         }
 
-        private Dictionary<string, string> GetScanProperties()
+        private static Dictionary<string, string> GetScanProperties(ICollection<Uri> discoveryUris, AuthenticationData authenticationData)
         {
             return new()
             {
-                { "NumberOfDiscoveryUris", _discoveryUris.Count.ToString() },
-                { "NumberOfUserCertificates", _authenticationData.userCertificates.Count.ToString() },
-                { "NumberOfAppCertificates", _authenticationData.applicationCertificates.Count.ToString() },
-                { "NumberOfLoginCredentials", _authenticationData.loginCredentials.Count.ToString() },
-                { "NumberOfBruteForceCredentials", _authenticationData.bruteForceCredentials.Count.ToString() },
+                { "NumberOfDiscoveryUris", discoveryUris.Count.ToString() },
+                { "NumberOfUserCertificates", authenticationData.userCertificates.Count.ToString() },
+                { "NumberOfAppCertificates", authenticationData.applicationCertificates.Count.ToString() },
+                { "NumberOfLoginCredentials", authenticationData.loginCredentials.Count.ToString() },
+                { "NumberOfBruteForceCredentials", authenticationData.bruteForceCredentials.Count.ToString() },
             };
         }
 
-        private Dictionary<string, string> GetScanResultProperties(Report report, TimeSpan timeSpan)
+        private static Dictionary<string, string> GetScanResultProperties(Report report, TimeSpan timeSpan, ICollection<Uri> discoveryUris, AuthenticationData authenticationData)
         {
             Dictionary<string, string> results = new()
             {
@@ -103,7 +89,7 @@ namespace Controller
                 { "NumberOfIssues",  report.Targets.Sum(t => t.IssuesCount).ToString() },
                 { "NumberOfErrors",  report.Targets.Sum(t => t.ErrorsCount).ToString() },
             };
-            GetScanProperties().ToList().ForEach(x => results.Add(x.Key, x.Value));
+            GetScanProperties(discoveryUris, authenticationData).ToList().ForEach(x => results.Add(x.Key, x.Value));
 
             return results;
         }
